abstract and organize seed helper functions
